Finish routes for billboards

+++ app/api/stores/[storeId]/route.ts
@@ -46,6 +46,10 @@ export async function DELETE(
 			return new NextResponse("Unauthorized", { status: 401 });
 		}
 
+		if (!params.storeId) {
+			return new NextResponse("Store id is required", { status: 400 });
+		}
+
 		const store = await prismadb.store.deleteMany({
 			where: {
 				id: params.storeId,
